+++ src/main/java/de/samply/exporter/focus/FocusQueryMetadata.java
@@ -1,0 +1,14 @@
+package de.samply.exporter.focus;
+
+import com.fasterxml.jackson.annotation.JsonProperty;
+import lombok.Data;
+
+@Data
+public class FocusQueryMetadata {
+
+    @JsonProperty("project")
+    private String project;
+    @JsonProperty("execute")
+    private boolean execute;
+
+}